[ALPS06702793] ffa: tmem: fix the ff-a handling
Fix the query function for ffa_handle to pa

MTK-Commit-Id: b2d3e056480b84f5196c641c3d2a7babcefc46db

CR-Id: ALPS06702793
Feature: devmpu

+++ drivers/tee/gud/510/TlcTui/tui-hal_mt.c
@@ -210,7 +210,7 @@ uint32_t hal_tui_alloc(
 {
 	uint32_t ret = TUI_DCI_ERR_INTERNAL_ERROR;
 	uint64_t pa = 0;
-	u32 sec_handle = 0;
+	u64 sec_handle = 0;
 	struct dma_heap *dma_heap;
 
 	if (!allocbuffer) {
